requirements.txt
asgiref==3.8.1
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
redis==5.0.7
sqlparse==0.5.0
tzdata==2024.1
python-decouple==3.8
PyMySQL==1.1.0
mysqlclient==2.2.4